API-18379 - introduce a ProductionRequest helper module
